// src/app/feature/dragons/models/select-options.model.ts
export interface SelectOption<T> {
	value: T;
	label: string;
	image?: string;
	icon?: string;
}

export interface ElementOption {
	value: string;
	label: string;
	image: string;
}

export interface SkillOption {
	value: string;
	label: string;
	image: string;
}

export interface RarityOption {
	value: string;
	label: string;
	icon: string;
}

export interface FamilyOption {
	value: string;
	label: string;
	icon: string;
}